% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{available_years}
\alias{available_years}
\title{Get available years}
\usage{
available_years(kpi, municipality = NULL, municipality_type = NULL)
}
\arguments{
\item{kpi}{One or more Kolada KPI IDs.}

\item{municipality}{Optional municipality or region name/code.}

\item{municipality_type}{Optional area type:
\code{"K"} = municipality, \code{"R"} = region.}
}
\value{
A data.frame with available years.
}
\description{
English wrapper for \code{\link[=tillgangliga_ar]{tillgangliga_ar()}}.
}
\examples{
\dontrun{
available_years(
  kpi = "N01926",
  municipality = "Haninge"
)
}

}
