% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{forandring}
\alias{forandring}
\title{Beräkna förändring mellan två år}
\usage{
forandring(nyckeltal, kommun = NULL, fran, till, kon = NULL, kommuntyp = NULL)
}
\arguments{
\item{nyckeltal}{Ett eller flera nyckeltals-ID från Kolada,
exempelvis \code{"N01926"} eller \code{c("N01926", "N17454")}.}

\item{kommun}{Valfritt. En eller flera kommunkoder eller
kommun-/regionnamn, exempelvis \code{"0136"} eller \code{"Haninge"}.
Om NULL används alla områden.}

\item{fran}{Det första året i jämförelsen.}

\item{till}{Det sista året i jämförelsen.}

\item{kon}{Valfritt filter för kön.
\code{"T"} = total, \code{"K"} = kvinnor och \code{"M"} = män.
Om NULL används alla tillgängliga kön.}

\item{kommuntyp}{Valfritt filter för områdestyp.
\code{"K"} = kommun och \code{"R"} = region.}
}
\value{
En data.frame med värde för startår och slutår,
absolut förändring och procentuell förändring.
}
\description{
Hämtar värden från Kolada för två valda år och beräknar
både absolut och procentuell förändring.
}
\details{
Funktionen kan användas för en eller flera kommuner,
regioner och nyckeltal.
}
\examples{
\dontrun{

forandring(
  nyckeltal = "N01926",
  kommun = "Haninge",
  fran = 2020,
  till = 2025,
  kon = "T"
)

forandring(
  nyckeltal = "N01926",
  kommuntyp = "K",
  fran = 2020,
  till = 2025,
  kon = "T"
)

}

}
