% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{hamta_enheter}
\alias{hamta_enheter}
\title{Hämta enheter från Kolada}
\usage{
hamta_enheter(
  verksamhet = NULL,
  kommun = NULL,
  nyckeltal = NULL,
  ar = NULL,
  kon = NULL,
  sok = NULL,
  batch_size = 25
)
}
\arguments{
\item{verksamhet}{Verksamhet eller V-kod, exempelvis \code{"Förskola"} eller \code{"V11"}.}

\item{kommun}{Valfri kommun angiven med namn eller kod.}

\item{nyckeltal}{Valfritt nyckeltals-ID.}

\item{ar}{Valfritt år eller flera år.}

\item{kon}{Valfritt kön: \code{"T"}, \code{"K"} eller \code{"M"}.}

\item{sok}{Valfri söktext för enhetsnamn.}

\item{batch_size}{Antal enheter per API-anrop. Standard är 25.}
}
\value{
En data.frame med enheter och, om nyckeltal anges,
nyckeltalsvärden på enhetsnivå.
}
\description{
Hämtar organisatoriska enheter och, valfritt,
nyckeltalsdata på enhetsnivå.
}
\details{
Följande verksamheter kan anges med namn eller V-kod:

\tabular{ll}{
\strong{Verksamhet} \tab \strong{V-kod} \cr
Förskola \tab V11 \cr
Grundskola F-9 \tab V15 \cr
Gymnasieskola \tab V17 \cr
Hemtjänst, äldre \tab V21 \cr
Särskilt boende, äldre \tab V23 \cr
LSS boende med särskild service \tab V25 \cr
LSS daglig verksamhet \tab V26 \cr
Gruppbostad LSS \tab V29 \cr
Servicebostad LSS \tab V30 \cr
SoL boendestöd \tab V31 \cr
SoL boende med särskild service \tab V32 \cr
SoL sysselsättning \tab V34 \cr
Våld i nära relationer \tab V45 \cr
Fastigheter \tab V60
}

Verksamhet kan anges antingen med namn, exempelvis
\code{verksamhet = "Förskola"}, eller direkt med V-kod,
exempelvis \code{verksamhet = "V11"}.

Exakta värden som kan anges i \code{verksamhet} är:

\tabular{ll}{
\strong{verksamhet} \tab \strong{V-kod} \cr
förskola \tab V11 \cr
grundskola \tab V15 \cr
gymnasieskola \tab V17 \cr
hemtjänst \tab V21 \cr
särskilt boende \tab V23 \cr
lss boende med särskild service \tab V25 \cr
lss daglig verksamhet \tab V26 \cr
gruppbostad lss \tab V29 \cr
servicebostad lss \tab V30 \cr
sol boendestöd \tab V31 \cr
sol boende med särskild service \tab V32 \cr
sol sysselsättning \tab V34 \cr
våld i nära relationer \tab V45 \cr
fastigheter \tab V60
}
}
\examples{
\dontrun{
# Hämta alla förskolor i Haninge
hamta_enheter(
  verksamhet = "Förskola",
  kommun = "Haninge"
)

# Samma sak med V-kod och kommunkod
hamta_enheter(
  verksamhet = "V11",
  kommun = "0136"
)

# Hämta även nyckeltalsvärden
hamta_enheter(
  verksamhet = "V11",
  kommun = "0136",
  nyckeltal = "N11808",
  ar = 2025
)

#' # Hämta nyckelord
hamta_enheter(
  verksamhet = "V11",
  nyckeltal = "N11808",
  ar = 2025,
  kommun = "Montessori"
)
}

}
