% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{info_nyckeltal}
\alias{info_nyckeltal}
\title{Visa information om nyckeltal}
\usage{
info_nyckeltal(nyckeltal)
}
\arguments{
\item{nyckeltal}{Ett eller flera nyckeltals-ID,
exempelvis \code{"N01926"} eller
\code{c("N01926", "N17454")}.}
}
\value{
En data.frame med metadata om nyckeltalen.
}
\description{
Hämtar metadata för ett eller flera nyckeltal från Kolada.
}
\examples{
\dontrun{
info_nyckeltal("N01926")

info_nyckeltal(
  c("N01926", "N17454")
)
}

}
