% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\docType{package}
\name{koladafun-package}
\alias{koladafun}
\alias{koladafun-package}
\title{koladafun: Functions for Kolada}
\description{
\code{koladafun} is an R package with functions for searching,
exploring and retrieving data from Kolada.
}
\details{
Paketet kan användas med både svenska och engelska funktionsnamn.
}
\section{Svenska funktioner}{

\itemize{
\item \code{\link[=hamta_fran_kolada]{hamta_fran_kolada()}} - Hämta data från Kolada.
\item \code{\link[=sok_nyckeltal]{sok_nyckeltal()}} - Sök efter nyckeltal.
\item \code{\link[=hamta_kommuner]{hamta_kommuner()}} - Hämta kommuner och regioner.
\item \code{\link[=tillgangliga_ar]{tillgangliga_ar()}} - Visa tillgängliga år.
\item \code{\link[=senaste_varde]{senaste_varde()}} - Hämta senaste tillgängliga värde.
\item \code{\link[=forandring]{forandring()}} - Beräkna förändring mellan två år.
\item \code{\link[=hamta_jamforelse]{hamta_jamforelse()}} - Jämför en kommun med andra kommuner.
\item \code{\link[=hamta_enheter]{hamta_enheter()}} - Hämta enheter och enhetsdata.
}
}

\section{English functions}{

\itemize{
\item \code{\link[=get_from_kolada]{get_from_kolada()}} - Get data from Kolada.
\item \code{\link[=search_kpi]{search_kpi()}} - Search for KPIs.
\item \code{\link[=get_municipalities]{get_municipalities()}} - Get municipalities and regions.
\item \code{\link[=available_years]{available_years()}} - Get available years.
\item \code{\link[=latest_value]{latest_value()}} - Get the latest available value.
\item \code{\link[=change]{change()}} - Calculate change between two years.
\item \code{\link[=get_comparison]{get_comparison()}} - Compare a municipality with other municipalities.
\item \code{\link[=get_units]{get_units()}} - Get organizational units and unit-level data.
}
}

\section{More information}{


README and source code are available on GitHub:
\url{https://github.com/RKAkolada/koladafun}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RKAKolada/koladafun}
  \item Report bugs at \url{https://github.com/RKAKolada/koladafun/issues}
}

}
\author{
\strong{Maintainer}: Nejdet Eken \email{nejdet.eken@rka.nu}

Authors:
\itemize{
  \item Nejdet Eken \email{nejdet.eken@rka.nu}
}

}
