% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{kpi_info}
\alias{kpi_info}
\title{Get information about KPIs}
\usage{
kpi_info(kpi)
}
\arguments{
\item{kpi}{One or more Kolada KPI IDs.}
}
\value{
A data.frame with KPI metadata.
}
\description{
English wrapper for \code{\link[=info_nyckeltal]{info_nyckeltal()}}.
}
\examples{
\dontrun{
kpi_info("N01926")
kpi_info(c("N01926", "N17454"))
}

}
