% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{latest_value}
\alias{latest_value}
\title{Get latest available value}
\usage{
latest_value(
  kpi,
  municipality = NULL,
  gender = NULL,
  municipality_type = NULL,
  include_missing = FALSE
)
}
\arguments{
\item{kpi}{One or more Kolada KPI IDs.}

\item{municipality}{Optional municipality or region.}

\item{gender}{Optional gender filter.}

\item{municipality_type}{Optional area type:
\code{"K"} = municipality, \code{"R"} or \code{"L"} = region.}

\item{include_missing}{Logical. If \code{FALSE}, the latest observation
with an actual value is returned. If \code{TRUE}, observations with
missing values due to e.g. confidentiality or missing data are
also accepted. Default is \code{FALSE}.}
}
\value{
A data.frame with the latest available observation.
}
\description{
English wrapper for \code{\link[=senaste_varde]{senaste_varde()}}.
}
\examples{
\dontrun{
latest_value(
  kpi = "N01926",
  municipality = "Haninge"
)
}

}
