% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{senaste_varde}
\alias{senaste_varde}
\title{Hämta senaste tillgängliga värde}
\usage{
senaste_varde(
  nyckeltal,
  kommun = NULL,
  kon = NULL,
  kommuntyp = NULL,
  bortfall = FALSE
)
}
\arguments{
\item{nyckeltal}{Ett eller flera nyckeltals-ID.}

\item{kommun}{Valfri kommun eller region.}

\item{kon}{Valfritt kön: \code{"T"}, \code{"K"} eller \code{"M"}.}

\item{kommuntyp}{Valfri områdestyp:
\code{"K"} = kommun och \code{"R"} = region.}

\item{bortfall}{Logiskt värde. Om \code{FALSE} hämtas senaste
observationen där ett faktiskt värde finns. Om \code{TRUE}
accepteras även observationer där värdet saknas, exempelvis
på grund av bortfall eller sekretess. Standard är \code{FALSE}.}
}
\value{
En data.frame med senaste tillgängliga observation.
}
\description{
Hämtar det senast tillgängliga värdet för ett eller flera
nyckeltal och områden.
}
\details{
Det senaste året bestäms separat för varje kombination av
nyckeltal, område och kön.
}
\examples{
\dontrun{
senaste_varde(
  nyckeltal = "N01926",
  kommun = "Haninge"
)

senaste_varde(
  nyckeltal = "N01926",
  kommun = "Haninge",
  bortfall = TRUE
)
}

}
