% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{tillgangliga_ar}
\alias{tillgangliga_ar}
\title{Visa tillgängliga år för nyckeltal}
\usage{
tillgangliga_ar(nyckeltal, kommun = NULL, kommuntyp = NULL)
}
\arguments{
\item{nyckeltal}{Ett eller flera nyckeltals-ID.}

\item{kommun}{Valfri kommun eller region, angiven som
namn eller kod.}

\item{kommuntyp}{Valfritt filter för områdestyp.
\code{"K"} = kommun och \code{"R"} eller \code{"L"} = region.
Koladas API använder \code{"L"} för region, men både \code{"R"} och \code{"L"}
accepteras av funktionen.
Flera kan anges med exempelvis \code{c("K", "R")}.}
}
\value{
En data.frame med nyckeltals-ID och tillgängliga år.
}
\description{
Hämtar vilka år som har data för ett eller flera
nyckeltal i Kolada.
}
\examples{
\dontrun{
tillgangliga_ar("N01926")

tillgangliga_ar(
  "N01926",
  kommun = "Haninge"
)
}

}
