% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{get_municipalities}
\alias{get_municipalities}
\title{Get municipalities and regions}
\usage{
get_municipalities(type = NULL, search = NULL)
}
\arguments{
\item{type}{Optional area type:
\code{"K"} = municipality, \code{"R"} = region.}

\item{search}{Optional search text for name or code.}
}
\value{
A data.frame with municipalities and regions.
}
\description{
English wrapper for \code{\link[=hamta_kommuner]{hamta_kommuner()}}.
}
\examples{
\dontrun{
get_municipalities()
get_municipalities(type = "K")
get_municipalities(search = "Han")
}

}
