% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{hamta_jamforelse}
\alias{hamta_jamforelse}
\title{Hämta jämförelse för kommun}
\usage{
hamta_jamforelse(nyckeltal, kommun, ar, kon = NULL)
}
\arguments{
\item{nyckeltal}{Ett eller flera nyckeltals-ID från Kolada.}

\item{kommun}{En kommun angiven med kommunkod eller kommunnamn.}

\item{ar}{Ett eller flera år.}

\item{kon}{Valfritt kön: \code{"T"}, \code{"K"} eller \code{"M"}.}
}
\value{
En data.frame med värden för kommunen och länets kommungrupp.
}
\description{
Hämtar data för en vald kommun och jämför med
länets kommungrupp (ovägt medel).
}
\examples{
\dontrun{
hamta_jamforelse(
  nyckeltal = "N01926",
  kommun = "Haninge",
  ar = 2025,
  kon = "T"
)
}

}
