% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{hamta_kommuner}
\alias{hamta_kommuner}
\title{Hämta kommuner och regioner från Kolada}
\usage{
hamta_kommuner(typ = NULL, sok = NULL)
}
\arguments{
\item{typ}{Valfri områdestyp.
\code{"K"} = kommun och \code{"R"} eller \code{"L"} = region.}

\item{sok}{Valfri söktext för namn eller kod.}
}
\value{
En data.frame med områdeskod, namn och typ.
}
\description{
Hämtar en lista över kommuner och regioner från Kolada.
Resultatet kan filtreras på områdestyp eller namn/kod.
}
\examples{
\dontrun{
hamta_kommuner()
hamta_kommuner(typ = "K")
hamta_kommuner(typ = "R")
hamta_kommuner(sok = "Han")
}

}
