% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\docType{package}
\name{koladafun}
\alias{koladafun-package}
\alias{koladafun}
\title{koladafun: Funktioner för Kolada}
\description{
\code{koladafun} är ett R-paket med funktioner för att söka,
utforska och hämta data från Kolada.
}
\section{Funktioner}{

\itemize{
\item \code{\link[=hamta_fran_kolada]{hamta_fran_kolada()}} - Hämta data från Kolada.
\item \code{\link[=sok_nyckeltal]{sok_nyckeltal()}} - Sök efter nyckeltal.
\item \code{\link[=info_nyckeltal]{info_nyckeltal()}} - Visa information om nyckeltal.
\item \code{\link[=hamta_kommuner]{hamta_kommuner()}} - Hämta kommuner och regioner.
\item \code{\link[=tillgangliga_ar]{tillgangliga_ar()}} - Visa tillgängliga år.
\item \code{\link[=senaste_varde]{senaste_varde()}} - Hämta senaste tillgängliga värde.
\item \code{\link[=forandring]{forandring()}} - Beräkna förändring mellan två år.
\item \code{\link[=hamta_jamforelse]{hamta_jamforelse()}} - Jämför en kommun med länets kommuner.
}
}

\section{Mer information}{


Paketets README och källkod finns på GitHub:
\url{https://github.com/RKAkolada/koladafun}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RKAKolada/koladafun}
  \item Report bugs at \url{https://github.com/RKAKolada/koladafun/issues}
}

}
\author{
\strong{Maintainer}: Nejdet Eken \email{nejdet.eken@rka.nu}

Authors:
\itemize{
  \item Nejdet Eken \email{nejdet.eken@rka.nu}
}

}
