% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{search_kpi}
\alias{search_kpi}
\title{Search for KPIs in Kolada}
\usage{
search_kpi(query, max_results = 50)
}
\arguments{
\item{query}{Search term or words.}

\item{max_results}{Maximum number of results. Default is 50.}
}
\value{
A data.frame with matching KPIs.
}
\description{
English wrapper for \code{\link[=sok_nyckeltal]{sok_nyckeltal()}}.
}
\examples{
\dontrun{
search_kpi("preschool")
search_kpi("N01926")
}

}
