% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Koladafun.R
\name{sok_nyckeltal}
\alias{sok_nyckeltal}
\title{Sök efter nyckeltal i Kolada}
\usage{
sok_nyckeltal(sok = NULL)
}
\arguments{
\item{sok}{Valfritt. Ett eller flera sökord eller
nyckeltals-ID. Om NULL returneras alla nyckeltal.}
}
\value{
En data.frame med nyckeltalsmetadata.
}
\description{
Hämtar metadata för Koladas nyckeltal.
Om inget sökord anges returneras samtliga nyckeltal.
}
\details{
Sökningen görs i nyckeltals-ID, namn och beskrivning.
Flera ord i samma sökfras kombineras med OCH.
Flera sökfraser kombineras med ELLER.
}
\examples{
\dontrun{
sok_nyckeltal()
sok_nyckeltal("förskola")
sok_nyckeltal("kostnad förskola")
sok_nyckeltal(c("N01926", "N17448"))
}

}
